requirements.txt
# Core FastAPI dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.8.0
pydantic-settings==2.1.0

# Database dependencies
asyncpg==0.29.0
SQLAlchemy==2.0.29
alembic==1.13.1

# Redis dependencies
redis==5.0.3
aioredis==2.0.1

# HTTP client dependencies
httpx==0.27.0
aiohttp==3.9.1

# AI service dependencies
openai==1.30.0
groq==0.4.1

# WhatsApp and messaging
twilio==8.10.3
requests==2.31.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Template engine
jinja2==3.1.2

# Environment and configuration
python-dotenv==1.0.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Date and time utilities
python-dateutil==2.8.2

# Validation and utilities
email-validator==2.1.0
phonenumbers==8.13.26

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Payment processing
razorpay==1.3.0

# Language processing
googletrans==4.0.0-rc1
langdetect==1.0.9

# File handling
python-magic==0.4.27
pillow==10.1.0

# Async utilities
asyncio-mqtt==0.16.1
websockets==12.0

# Optional tooling
watchdog==4.0.0          # used by start-dev.sh for auto-reload
